server/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
Appium-Python-Client==1.0.2
astroid==2.4.2
attrs==19.3.0
axe-selenium-python==2.1.6
blinker==1.4
certifi==2020.6.20
cffi==1.14.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
cryptography==3.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
flask-expects-json==1.4.0
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
identify==1.4.25
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
nodeenv==1.4.0
packaging==20.4
passlib==1.7.2
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
selenium==3.141.0
sentry-sdk==0.16.3
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
